supporting large inputs

+++ src/shogun/lib/common.h
@@ -69,7 +69,7 @@ typedef char                       string_t[STRING_LEN];
 typedef int                        machine_int_t;
 
 /** index type */
-typedef int32_t index_t;
+typedef int64_t index_t;
 
 /** complex type */
 #include <complex>
